Clarify that manual configuration of rover realm may not be needed
When using the default helm chart, the rover realm and the rover client will automatically be created on first start of Iris. 

Therefore those things should not be configured by humans manually. 

This update hints at the fact that this manual configuration step should be skipped if the iris helm chart is used.